- gcc 4.0 warning & fixes "forward" ported from stable - postgres fixes/warnings "forward" ported from stable

+++ modules_s/acc/diam_avp.c
@@ -105,7 +105,7 @@ AAA_AVP*  AAACreateAVP(
 	AAA_AVPFlag flags,
 	AAAVendorId vendorId,
 	char   *data,
-	unsigned int length,
+	size_t length,
 	AVPDataStatus data_status)
 {
 	AAA_AVP *avp;
